Further refactoring.
